+++ Sources/Internal/Render/2D/GraphicsFont.h
@@ -48,7 +48,7 @@ public:
 	static GraphicsFont * Create(const FilePath & fontDefName, const FilePath & spriteName);
 	
 	virtual void	SetSize(float32 size);
-	virtual Size2i	GetStringSize(const WideString & str, Vector<int32> *charSizes = NULL) const;
+	virtual Size2i	GetStringSize(const WideString & str, Vector<float32> *charSizes = NULL) const;
 	virtual bool	IsCharAvaliable(char16 ch) const;
 	virtual uint32	GetFontHeight() const;
     virtual int32   GetHorizontalSpacing() const;
